+++ src/pages/GeneralPage/GeneralPage.tsx
@@ -43,6 +43,7 @@ export default function GeneralPage() {
   const handleCloseAbstractLatexModal = () => setOpenAbstractLatexModal(false);
 
   React.useEffect(() => {
+    console.log('TREVOR', getProposal());
     setValidateToggle(!validateToggle);
     helpComponent(t('abstract.help'));
   }, []);